fix: changed helpers package name and fixed cli params loading
