+++ src/main/kotlin/com/enniovisco/Main.kt
@@ -11,8 +11,8 @@ import java.net.*
 private typealias ResultData = List<Map<String, String>>
 private typealias Metadata = Map<String, String>
 
-fun main(args: Array<String>) {
-    Cli(args, toFile = true) {
+fun main(args: Array<String>, preloaded: Boolean = false) {
+    Cli(args, toFile = true, preloaded = preloaded) {
         it.title("Tracking")
         val snapshots = tracking(it)
