+++ src/main/kotlin/com/enniovisco/Shared.kt
@@ -99,23 +99,3 @@ object Spec {
      */
     val screen = AtomicFormula("screen")
 }
-
-class WebMonitor {
-    fun webSource(init: WebSource.() -> Unit): WebSource {
-        WebSource.init()
-        return WebSource
-    }
-
-    fun spec(init: Spec.() -> Unit): Spec {
-        Spec.init()
-        return Spec
-    }
-}
-
-fun monitor(init: WebMonitor.()-> Unit) {
-    val webmonitor = WebMonitor()
-    webmonitor.init()
-    println("Starting WebMonitor...")
-    main(emptyArray())
-}
-
