+++ src/main/kotlin/com/enniovisco/cli/Cli.kt
@@ -7,6 +7,7 @@ class Cli(
     args: Array<String>,
     toConsole: Boolean = false,
     toFile: Boolean = false,
+    preloaded: Boolean = false,
     exec: Cli.(Reporter) -> Unit
 ) {
     private val report = Reporter(toConsole, toFile)
@@ -14,7 +15,9 @@ class Cli(
 
     init {
         report.use {
-            validateArgs(args)
+            if(!preloaded) {
+                validateArgs(args)
+            }
             exec(it)
         }
     }
